+++ assets/js/components/channels/ChannelNew.jsx
@@ -2,23 +2,23 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import ChannelDashboardLayout from "./ChannelDashboardLayout";
-import AzureHubForm from "./forms/AzureHubForm.jsx";
-import AzureCentralForm from "./forms/AzureCentralForm.jsx";
-import AWSForm from "./forms/AWSForm.jsx";
-import GoogleForm from "./forms/GoogleForm.jsx";
-import MQTTForm from "./forms/MQTTForm.jsx";
-import HTTPForm from "./forms/HTTPForm.jsx";
-import CargoForm from "./forms/CargoForm.jsx";
-import MyDevicesForm from "./forms/MyDevicesForm.jsx";
-import AdafruitForm from "./forms/AdafruitForm.jsx";
-import UbidotsForm from "./forms/UbidotsForm.jsx";
-import DatacakeForm from "./forms/DatacakeForm.jsx";
-import TagoForm from "./forms/TagoForm.jsx";
-import GoogleSheetForm from "./forms/GoogleSheetForm.jsx";
-import MicroshareForm from "./forms/MicroshareForm.jsx";
+import AzureHubForm from "./default/AzureHubForm.jsx";
+import AzureCentralForm from "./default/AzureCentralForm.jsx";
+import AWSForm from "./default/AWSForm.jsx";
+import GoogleForm from "./default/GoogleForm.jsx";
+import MQTTForm from "./default/MQTTForm.jsx";
+import HTTPForm from "./default/HTTPForm.jsx";
+import CargoForm from "./community/cargo/CargoForm.jsx";
+import MyDevicesForm from "./community/my_devices/MyDevicesForm.jsx";
+import AdafruitForm from "./community/adafruit/AdafruitForm.jsx";
+import UbidotsForm from "./community/ubidots/UbidotsForm.jsx";
+import DatacakeForm from "./community/datacake/DatacakeForm.jsx";
+import TagoForm from "./community/tago/TagoForm.jsx";
+import GoogleSheetForm from "./community/google_sheets/GoogleSheetForm.jsx";
+import MicroshareForm from "./community/microshare/MicroshareForm.jsx";
+import AkenzaForm from "./community/akenza/AkenzaForm";
 import ChannelCreateRow from "./ChannelCreateRow";
 import ChannelPremadeRow from "./ChannelPremadeRow";
-import AkenzaForm from "./forms/AkenzaForm";
 import { MobileDisplay, DesktopDisplay } from "../mobile/MediaQuery";
 import { createChannel } from "../../actions/channel";
 import analyticsLogger from "../../util/analyticsLogger";
