+++ assets/js/components/channels/community/cargo/CargoForm.jsx
@@ -1,20 +1,24 @@
 import React, { Component } from 'react';
-import { IntegrationTypeTileSimple } from "../IntegrationTypeTileSimple";
-import { getRootType } from "../../../util/integrationInfo";
+import { IntegrationTypeTileSimple } from "../../IntegrationTypeTileSimple";
+import { getRootType } from "../../../../util/integrationInfo";
 import { Link } from "react-router-dom";
-import ChannelNameForm from "./ChannelNameForm.jsx";
-import analyticsLogger from "../../../util/analyticsLogger";
+import ChannelNameForm from "../../default/ChannelNameForm.jsx";
+import analyticsLogger from "../../../../util/analyticsLogger";
 import { Card, Typography, Button } from 'antd';
 const { Text } = Typography
 
-class MyDevicesForm extends Component {
+class CargoForm extends Component {
   state = {
     method: "post",
-    endpoint: "https://lora.mydevices.com/v1/networks/helium/uplink",
-    headers: {},
-    showNextSteps: true,
-    validInput: true,
+    endpoint: "https://cargo.helium.com/api/payloads",
+    headers: { "Content-Type": "application/json" },
+    showNextSteps: false,
+    validInput: false,
     channelName: "",
+  }
+
+  validateInput = () => {
+    this.setState({ showNextSteps: true, validInput: true })
   }
 
   handleNameInput = (e) => {
@@ -67,8 +71,24 @@ class MyDevicesForm extends Component {
 
         <Card title="Step 2 - Endpoint Details">
           <Text>
-            Use a supported device listed on their console or encode the payload with the Cayenne Low Power Payload (Cayenne LPP) format.
+            You are opting to test your devices on the Helium Cargo HTTP integration endpoint.
           </Text>
+          <br />
+          <Text>
+            Helium Cargo is an evaluation tool and the data collected is available to all developers.
+          </Text>
+          <br />
+          <Text>
+            Please do not share any sensitive information. Use at your own discretion.
+          </Text>
+          <br />
+          <Button
+            onClick={this.validateInput}
+            type="default"
+            style={{marginTop: 20}}
+          >
+            I AGREE
+          </Button>
         </Card>
 
         {this.state.showNextSteps && (
@@ -85,4 +105,4 @@ class MyDevicesForm extends Component {
   }
 }
 
-export default MyDevicesForm;
+export default CargoForm;
